ai_engine/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# CrewAI (Phase-1 framework)
crewai==0.11.0

# LLM provider: Gemini (recommended default)
langchain-google-genai==1.0.3

# Optional provider: OpenAI
langchain-openai==0.0.8

# Other utils
psutil==5.9.8
